refactor: consolidate rebase_authorship.rs into rewrite_op_v3.rs
Move all functions from rebase_authorship.rs into the unified
rewrite_op_v3.rs module, update all callers to import from the
new location, and delete the old file along with 12 orphaned
snapshot files.

Net reduction of 745 lines. No logic changes — pure code movement
and import path updates. All 4622 tests pass.
